[P-13] fix personal property load logic

+++ app/src/main/java/com/marshall/pyerite/personalPropertyModule/model/PersonalPropertyModels.kt
@@ -1,7 +1,7 @@
 package com.marshall.pyerite.personalPropertyModule.model
 
 internal data class PersonalPropertyBucket(
-    val count: Int? = null,
+    val count: Long? = null,
     val isk: Double? = null,
 )
